fix: Aktivitäten werden jetzt nur von Membern mit Team zurückgegeben
Refs: SCP-157
Time: 0.3 h

+++ Backend/src/test/java/de/hsesslingen/scpprojekt/scp/Database/Services/MemberServiceTest.java
@@ -73,7 +73,7 @@ public class MemberServiceTest {
             memberList.add(m);
 
             when(memberRepository.findMemberByEmail("test" + i + "@example.com")).thenReturn(m);
-            when(memberRepository.findById((long)i)).thenReturn(Optional.of(m));
+            when(memberRepository.findById(i)).thenReturn(Optional.of(m));
         }
 
         when(memberRepository.findAll()).thenReturn(memberList);
